display upload & success state
